adjustments to schedule evaluation graph

+++ django_server_files/core/management/commands/remove_concrete_pulsar_from_db.py
@@ -2,11 +2,19 @@ from django.core.management.base import BaseCommand
 from core.models import Pulsar
 
 class Command(BaseCommand):
-    help = 'Removes specific pulsars objects from the database'
+    help = 'Removes specific pulsar objects from the database'
+
+    def add_arguments(self, parser):
+        # Add argument to specify pulsar name
+        parser.add_argument('pulsar_name', type=str, help='The name of the pulsar to be removed')
 
     def handle(self, *args, **options):
-        remove_concrete_pulsar(args[1])
+        pulsar_name = options['pulsar_name']
+        remove_concrete_pulsar(pulsar_name)
 
 def remove_concrete_pulsar(name):
-    Pulsar.objects.filter(name=name).delete()
-    print(f"pulsar {name} was removed")
+    if Pulsar.objects.filter(name=name).exists():
+        Pulsar.objects.filter(name=name).delete()
+        print(f"Pulsar {name} was removed")
+    else:
+        print(f"No pulsar named {name} found")
